requirements.txt
datasets==3.3.1
anthropic==0.34.2
python-dotenv==1.0.1
transformers>=4.22.0
torch>=1.10.0
tqdm>=4.63.0
numpy>=1.22.0
